custom dev server
We've been using esbuild as dev server so far, but the problem with that is that it  serves audio files with "application/octet-stream" content-type and Chrome does not let you seek the audio then. This commit puts a proxy in front of esbuild server to server audio files with correct types.

skip-checks:true